Add Repository Links (#43)
# Pull Request

<!-- Please use British English for spelling and terminology where
possible. However, if you're more comfortable using another variant
don't worry about it! -->

## Description:

This pull request includes updates to the `IDEAS.md` file to standardise
the format of GitHub repository links. The changes ensure consistency
across the document by modifying the link text and structure.

Standardisation of GitHub repository links:

*
[`IDEAS.md`](diffhunk://#diff-79b8416fcbe9d355cc6dcda13b825ff89aa55496709be20e0c5621ad6a58417aR61-R62):
Updated the link format for the GitHub repository associated with the
action that analyses repositories.
*
[`IDEAS.md`](diffhunk://#diff-79b8416fcbe9d355cc6dcda13b825ff89aa55496709be20e0c5621ad6a58417aR71-R72):
Updated the link format for the GitHub repository associated with the
Chrome extension for spell-checking code.
*
[`IDEAS.md`](diffhunk://#diff-79b8416fcbe9d355cc6dcda13b825ff89aa55496709be20e0c5621ad6a58417aL77-R81):
Standardized the link format for the GitHub repository associated with
the action that analyses pull requests.
*
[`IDEAS.md`](diffhunk://#diff-79b8416fcbe9d355cc6dcda13b825ff89aa55496709be20e0c5621ad6a58417aL124-R129):
Modified the link text and structure for the GitHub repository and
GitHub Pages associated with the Tech Radar project.
*
[`IDEAS.md`](diffhunk://#diff-79b8416fcbe9d355cc6dcda13b825ff89aa55496709be20e0c5621ad6a58417aL137-R141):
Updated the link format for the GitHub repository associated with the
screenshotting of emails on Mailinator.