Update Pull Request Template (#44)
# Pull Request

<!-- Please use British English for spelling and terminology where
possible. However, if you're more comfortable using another variant
don't worry about it! -->

## Description

This pull request includes a modification to the
`.github/pull_request_template.md` file to improve the clarity and
detail of pull request descriptions.

Changes to pull request template:

*
[`.github/pull_request_template.md`](diffhunk://#diff-b2496e80299b8c3150b1944450bd81c622e04e13d15c411d291db0927d75fd6bL5-R7):
Replaced the `TL;DR` section with a `Description` section. Updated the
instructions to include a summary of the change, the issue it fixes,
relevant motivation and context, and any dependencies required for the
change.