Add Actionlint Job (#57)
# Pull Request

## Description

This pull request introduces a new job to the GitHub Actions workflow
and disables comments in the YAML linting configuration. The most
important changes include adding a job to check GitHub Actions with
`actionlint` and updating the YAML linting rules.

Enhancements to GitHub Actions workflow:

*
[`.github/workflows/check-everything.yml`](diffhunk://#diff-4dd84dd6da9ad3de39ce8fb6ebb289ee7a656956984dbe3ff166848b6602a171R49-R63):
Added a new job `run-actionlint` to check GitHub Actions with
`actionlint`, including steps to checkout the repository, install
`actionlint`, and run the checks.

Updates to YAML linting configuration:

*
[`.github/super-linter-configs/.yaml-lint.yml`](diffhunk://#diff-cdea8bea457c52a8c0651465a847dcbc52c6abd1fb370c531b89a3214efad8bbR7):
Disabled comments in the YAML linting rules.