Update Technologies (#64)
# Pull Request

## Description

This pull request includes updates to the workflow configuration and
documentation files. The primary changes are updating dependencies in
the workflow and modifying the tech stack in the documentation.

Workflow configuration updates:

*
[`.github/workflows/check-everything.yml`](diffhunk://#diff-4dd84dd6da9ad3de39ce8fb6ebb289ee7a656956984dbe3ff166848b6602a171L64-R72):
Updated the `astral-sh/setup-uv` action from version `v4.2.0` to
`v5.0.1` and the `github/codeql-action/upload-sarif` action from version
`v3.27.9` to `v3.28.0`.

Documentation updates:

*
[`IDEAS/STARTED_IDEAS.md`](diffhunk://#diff-1f625e6769ecd0974fe214e77a5a3a0b1702b21488d4b933d1d3c7ed2b2d5837L56-R57):
Removed `TypeScript` and `Astro` from the tech stack and added
`Markdown`.