Scan GitHub Actions with CodeQL (#69)
This pull request includes updates to the
`.github/workflows/check-everything.yml` file to enhance the CI/CD
pipeline by updating dependencies and adding a new CodeQL analysis job.

Dependency updates:

* Updated `astral-sh/setup-uv` from version `v5.0.1` to `v5.1.0` to
ensure the latest features and fixes are included.
* Updated `github/codeql-action/upload-sarif` from version `v3.28.0` to
`v3.28.1` to include the latest improvements and bug fixes.

New job addition:

* Added a new `run-codeql-analysis` job to perform CodeQL analysis,
which includes steps to checkout the repository, initialize CodeQL, and
perform the analysis. This enhances the security and quality checks in
the CI/CD pipeline.