Check Justfile Format (#84)
# Pull Request

## Description

This pull request includes updates to the
`.github/workflows/code-checks.yml` file to add a new job for checking
the Justfile format and to update the versions of several GitHub
Actions.

New job addition:

* Added a new job `check-justfile-format` to the GitHub Actions workflow
to ensure the Justfile format is correct. This job includes steps for
checking out the repository, setting up Just, and running the format
check.

Version updates:

* Updated the `astral-sh/setup-uv` action from version `v5.2.2` to
`v5.3.0` for installing the latest version of uv.
* Updated the `github/codeql-action/upload-sarif` action from version
`v3.28.8` to `v3.28.10` for uploading SARIF files.
* Updated the `github/codeql-action/init` action from version `v3.28.8`
to `v3.28.10` for initializing CodeQL.
* Updated the `github/codeql-action/analyze` action from version
`v3.28.8` to `v3.28.10` for performing CodeQL analysis.