Update Dependency Review (#85)
# Pull Request

## Description

This pull request includes updates to the
`.github/workflows/dependency-review.yml` file to enhance the Dependency
Review workflow. The most important changes include adding permissions
for pull requests and configuring the action to comment on the pull
request summary upon failure.

Enhancements to Dependency Review workflow:

*
[`.github/workflows/dependency-review.yml`](diffhunk://#diff-7cdd3ccec44c8ba176bdc3b9ef54c3f56aa210a1a4e2bb5f79d87b1e50314a18R14-R15):
Added `pull-requests: write` permission to allow the workflow to
interact with pull requests.
*
[`.github/workflows/dependency-review.yml`](diffhunk://#diff-7cdd3ccec44c8ba176bdc3b9ef54c3f56aa210a1a4e2bb5f79d87b1e50314a18R24-R25):
Configured the `actions/dependency-review-action` to comment on the pull
request summary if the dependency review fails.