Remove CodeQL Configuration (#111)
# Pull Request

## Description

This pull request removes a custom CodeQL configuration file and updates
the workflow to no longer reference it. The most important changes are
as follows:

### Removal of custom CodeQL configuration:

*
[`.github/other-configurations/codeql-config.yml`](diffhunk://#diff-687a0af9316072cbdab79cdb046aac181231f8b5f3b5502bc796a9d949ce65beL1-L3):
Deleted the query filter that excluded the `actions/unpinned-tag` query.

### Workflow update:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L126):
Removed the `config-file` parameter in the CodeQL workflow, which
previously pointed to the now-removed custom configuration file.