Remove `.editorconfig-checker.json` (#132)
# Pull Request

## Description

This pull request removes the configuration for the
`editorconfig-checker` tool and updates the related GitHub Actions
workflow to reflect this change.

### Removal of `editorconfig-checker` configuration:

*
[`.github/super-linter-configurations/.editorconfig-checker.json`](diffhunk://#diff-926aaab00380cdd79b8c5069e4eb6d6a7b0a730b2cfd8254e271be7844a0bd10L1-L18):
Deleted the configuration file for `editorconfig-checker`, which
included settings for verbosity, debugging, and various linting rules.

### Updates to GitHub Actions workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L34):
Removed the `EDITORCONFIG_FILE_NAME` environment variable, which is no
longer needed after the removal of the `editorconfig-checker`
configuration.