Update Reusable Workflows (#142)
# Pull Request

## Description


This pull request updates several GitHub Actions workflows to use the
latest versions of reusable workflows, adds a new permission to one
workflow, and removes a redundant command from the `Justfile`.
Additionally, it updates the minimum version requirement for Lefthook in
`lefthook.yml`.

### Workflow updates:
*
[`.github/workflows/clean-caches.yml`](diffhunk://#diff-d0394e4336a74cdfc1d4cff05d056b893ac7ff922eacf4448e104a754f386b8dL15-R15):
Updated the reusable workflow version for cache cleaning to
`v2025.06.06.01`.
* `.github/workflows/code-checks.yml`: 
- Updated the reusable workflow versions for code checks and CodeQL
analysis to `v2025.06.06.01`.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R39-R42)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L53-R54)
  - Added `actions: read` permission to the code checks workflow.
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfL15-R15):
Updated the reusable workflow version for pull request tasks to
`v2025.06.06.01`.
*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6L18-R18):
Updated the reusable workflow version for syncing labels to
`v2025.06.06.01`.

### Code cleanup:
*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cL49-L52):
Removed the `zizmor-check-sarif` command, which generated SARIF output,
as it is no longer needed.

### Lefthook configuration:
*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dL2-R2):
Increased the minimum Lefthook version requirement from `1.11.11` to
`1.11.13`.