Update actionlint Git Hook (#173)
# Pull Request

## Description

This pull request adds a new validation step for GitHub Actions
workflows using `actionlint` and integrates it into the pre-commit
hooks. The most important changes are as follows:

### Addition of `actionlint` validation:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR49-R56):
Introduced a new `actionlint-check` command to run `actionlint` for
validating GitHub Actions workflows.

*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dR23-R24):
Added `Actionlint Checks` to the pre-commit hooks, which invokes the
`actionlint-check` command defined in the `Justfile`.